% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{exploregraph}
\alias{exploregraph}
\title{Convenience wrapper for \code{explore} for adjacency matrices}
\usage{
exploregraph(data, B, ...)
}
\arguments{
\item{data}{List of symmetric matrices, one matrix for each penalization
level}

\item{B}{Number of bootstraps used to construct \code{data}. At least 21 are
needed for u-shape test heuristic to work, but in general it is recommended
to use many more.}

\item{...}{Additional arguments are passed on to \code{explore}.}
}
\value{
A list with components
  \item{pop.sep}{vector of values saying how separated true and false
    variables are for each level of penalization}
}
\description{
When modeling graphs it may be more convenient to store data as matrices
instead of row vectors.
}
