% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{symmetric.matrix2vector}
\alias{symmetric.matrix2vector}
\title{Take upper half of matrix and convert it to a vector}
\usage{
symmetric.matrix2vector(m)
}
\arguments{
\item{m}{A symmetric matrix}
}
\description{
If variable selection counts are in a matrix this function converts them into
vector to input into rope. Can be useful when variables correspond to edges
in a graph.
}
