% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope.R
\name{vector2symmetric.matrix}
\alias{vector2symmetric.matrix}
\title{Convert vector that represents half of a symmetric matrix into a matrix}
\usage{
vector2symmetric.matrix(v)
}
\arguments{
\item{v}{A vector with length p*(p-1)/2 for some integer p}
}
\description{
This can be convenient for using output when rope is used for selection of
graph models.
}
